refactor(int64): core type fixes - NDArray, Storage, Iterator
- NDArray`1.cs: Add long[] indexer, int[] delegates to it
- UnmanagedStorage.cs: Add Span overflow check (Span limited to int)
- UnmanagedStorage.Cloning.cs: Add ArraySlice allocation overflow check
- NDIterator.cs: Change size field from int to long

Note: ~900 cascading errors remain from:
- ArraySlice (needs long count)
- Incrementors (need long coords)
- Various Default.* operations
- IKernelProvider interface

Part of int32 to int64 indexing migration (#584)

+++ src/NumSharp.Bitmap/np_.extensions.cs
@@ -284,7 +284,7 @@ namespace NumSharp
                             break;
                     }
 
-                    return bbp;
+                    return (int)bbp;
                 }
 
                 return format.ToBytesPerPixel();
